# openapi2.json
{"openapi":"3.1.0","info":{"title":"PM Weather API","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}},"/api/location/current":{"get":{"summary":"Current Location","operationId":"current_location_api_location_current_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Current Location Api Location Current Get"}}}}}}},"/api/locations/search":{"get":{"summary":"Search Locations","operationId":"search_locations_api_locations_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":2,"title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/weather/current":{"get":{"summary":"Current Weather","operationId":"current_weather_api_weather_current_get","parameters":[{"name":"city","in":"query","required":false,"schema":{"type":"string","minLength":0,"default":"","title":"City"}},{"name":"lat","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lat"}},{"name":"lon","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lon"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherSnapshot"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/weather/forecast":{"get":{"summary":"Forecast","operationId":"forecast_api_weather_forecast_get","parameters":[{"name":"city","in":"query","required":false,"schema":{"type":"string","minLength":0,"default":"","title":"City"}},{"name":"lat","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lat"}},{"name":"lon","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lon"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/weather/bundle":{"get":{"summary":"Bundle","operationId":"bundle_api_weather_bundle_get","parameters":[{"name":"city","in":"query","required":false,"schema":{"type":"string","minLength":0,"default":"","title":"City"}},{"name":"lat","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lat"}},{"name":"lon","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lon"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherBundle"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/videos/search":{"get":{"summary":"Videos Search","operationId":"videos_search_api_videos_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Q"}},{"name":"maxResults","in":"query","required":false,"schema":{"type":"integer","maximum":12,"minimum":1,"default":6,"title":"Maxresults"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/weather/history":{"post":{"summary":"Save Weather History Route","operationId":"save_weather_history_route_api_weather_history_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveWeatherHistoryRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherHistoryRecord"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Weather History Route","operationId":"list_weather_history_route_api_weather_history_get","parameters":[{"name":"location","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"}},{"name":"startDate","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Startdate"}},{"name":"endDate","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Enddate"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WeatherHistoryRecord"},"title":"Response List Weather History Route Api Weather History Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/weather/history/{record_id}":{"put":{"summary":"Update Weather History Route","operationId":"update_weather_history_route_api_weather_history__record_id__put","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveWeatherHistoryRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherHistoryRecord"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Weather History Route","operationId":"delete_weather_history_route_api_weather_history__record_id__delete","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Weather History Route","operationId":"get_weather_history_route_api_weather_history__record_id__get","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherHistoryRecord"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/weather/history/preview":{"post":{"summary":"Preview Weather History Route","operationId":"preview_weather_history_route_api_weather_history_preview_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveWeatherHistoryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherHistoryRecord"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ForecastDay":{"properties":{"day":{"type":"string","title":"Day"},"summary":{"type":"string","title":"Summary"},"lowC":{"type":"integer","title":"Lowc"},"highC":{"type":"integer","title":"Highc"},"icon":{"type":"string","title":"Icon"},"precipitationProbability":{"type":"integer","title":"Precipitationprobability"},"sunrise":{"type":"string","title":"Sunrise"},"sunset":{"type":"string","title":"Sunset"}},"type":"object","required":["day","summary","lowC","highC","icon","precipitationProbability","sunrise","sunset"],"title":"ForecastDay"},"ForecastHour":{"properties":{"time":{"type":"string","title":"Time"},"temperatureC":{"type":"integer","title":"Temperaturec"},"feelsLikeC":{"type":"integer","title":"Feelslikec"},"precipitationProbability":{"type":"integer","title":"Precipitationprobability"},"icon":{"type":"string","title":"Icon"},"summary":{"type":"string","title":"Summary"}},"type":"object","required":["time","temperatureC","feelsLikeC","precipitationProbability","icon","summary"],"title":"ForecastHour"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SaveWeatherHistoryRequest":{"properties":{"city":{"type":"string","minLength":0,"title":"City","default":""},"lat":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lat"},"lon":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lon"},"startDate":{"type":"string","format":"date","title":"Startdate"},"endDate":{"type":"string","format":"date","title":"Enddate"}},"type":"object","required":["startDate","endDate"],"title":"SaveWeatherHistoryRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WeatherBundle":{"properties":{"location":{"type":"string","title":"Location"},"current":{"$ref":"#/components/schemas/WeatherSnapshot"},"hourly":{"items":{"$ref":"#/components/schemas/ForecastHour"},"type":"array","title":"Hourly"},"daily":{"items":{"$ref":"#/components/schemas/ForecastDay"},"type":"array","title":"Daily"}},"type":"object","required":["location","current","hourly","daily"],"title":"WeatherBundle"},"WeatherHistoryRecord":{"properties":{"id":{"type":"integer","title":"Id"},"location":{"type":"string","title":"Location"},"city":{"type":"string","title":"City"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"lat":{"type":"number","title":"Lat"},"lon":{"type":"number","title":"Lon"},"startDate":{"type":"string","format":"date","title":"Startdate"},"endDate":{"type":"string","format":"date","title":"Enddate"},"savedAt":{"type":"string","title":"Savedat"},"bundle":{"$ref":"#/components/schemas/WeatherBundle"}},"type":"object","required":["id","location","city","lat","lon","startDate","endDate","savedAt","bundle"],"title":"WeatherHistoryRecord"},"WeatherSnapshot":{"properties":{"city":{"type":"string","title":"City"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"temperatureC":{"type":"integer","title":"Temperaturec"},"feelsLikeC":{"type":"integer","title":"Feelslikec"},"humidity":{"type":"integer","title":"Humidity"},"pressure":{"type":"integer","title":"Pressure"},"dewPointC":{"type":"integer","title":"Dewpointc"},"uvIndex":{"type":"number","title":"Uvindex"},"visibilityKm":{"type":"number","title":"Visibilitykm"},"windSpeedKmh":{"type":"number","title":"Windspeedkmh"},"windDirection":{"type":"string","title":"Winddirection"},"windGustKmh":{"type":"number","title":"Windgustkmh"},"sunrise":{"type":"string","title":"Sunrise"},"sunset":{"type":"string","title":"Sunset"},"condition":{"type":"string","title":"Condition"},"description":{"type":"string","title":"Description"},"updatedAt":{"type":"string","title":"Updatedat"}},"type":"object","required":["city","latitude","longitude","temperatureC","feelsLikeC","humidity","pressure","dewPointC","uvIndex","visibilityKm","windSpeedKmh","windDirection","windGustKmh","sunrise","sunset","condition","description","updatedAt"],"title":"WeatherSnapshot"}}}}